+++ src/collective/volto/formsupport/restapi/services/form_data/csv.py
@@ -2,8 +2,10 @@ from collective.volto.formsupport.interfaces import IFormDataStore
 from collective.volto.formsupport.utils import get_blocks
 from copy import deepcopy
 from io import StringIO
+from plone.namedfile import NamedBlobFile
 from plone.restapi.serializer.converters import json_compatible
 from plone.restapi.services import Service
+from zExceptions import NotFound
 from zope.component import getMultiAdapter
 
 import csv
@@ -16,12 +18,15 @@ class FormDataExportGet(Service):
     def __init__(self, context, request):
         super().__init__(context, request)
         self.form_fields_order = []
-        self.form_block = {}
+        self.form_block = None
+        self.block_id = self.request.get("block_id")
 
         blocks = getattr(context, "blocks", {})
         if not blocks:
-            return
+            raise NotFound("No blocks found")
         for id, block in blocks.items():
+            if self.block_id and id != self.block_id:
+                continue
             block_type = block.get("@type", "")
             if block_type == "form":
                 self.form_block = block
@@ -105,6 +110,8 @@ class FormDataExportGet(Service):
         fields_labels = self.get_fields_labels()
 
         for item in store.search():
+            if self.block_id and item.attrs.get("block_id") != self.block_id:
+                continue
             data = {}
 
             for k in self.get_ordered_keys(item):
@@ -119,7 +126,10 @@ class FormDataExportGet(Service):
                 if k not in self.form_fields_order and label not in legacy_columns:
                     legacy_columns.append(label)
 
-                data[label] = json_compatible(value)
+                if isinstance(value, NamedBlobFile):
+                    data[label] = value.filename
+                else:
+                    data[label] = json_compatible(value)
 
             for k in fixed_columns:
                 # add fixed columns values