+++ src/collective/volto/formsupport/tests/test_store_action_form.py
@@ -9,10 +9,13 @@ from plone.app.testing import SITE_OWNER_NAME
 from plone.app.testing import SITE_OWNER_PASSWORD
 from plone.app.testing import TEST_USER_ID
 from plone.restapi.testing import RelativeSession
+from plone.testing.zope import Browser
 from Products.MailHost.interfaces import IMailHost
 from zope.component import getUtility
 
+import base64
 import csv
+import os
 import transaction
 import unittest
 
@@ -24,6 +27,8 @@ class TestMailStore(unittest.TestCase):
         self.app = self.layer["app"]
         self.portal = self.layer["portal"]
         self.portal_url = self.portal.absolute_url()
+        self.browser = Browser(self.layer["app"])
+
         setRoles(self.portal, TEST_USER_ID, ["Manager"])
 
         self.mailhost = getUtility(IMailHost)
@@ -128,11 +133,21 @@ class TestMailStore(unittest.TestCase):
                         "field_id": "name",
                         "field_type": "text",
                     },
+                    {
+                        "label": "Attachment",
+                        "field_id": "file",
+                        "field_type": "attachment",
+                    },
                 ],
             },
         }
         transaction.commit()
 
+        filename = os.path.join(os.path.dirname(__file__), "file.pdf")
+        with open(filename, "rb") as f:
+            file_str = f.read()
+
+        # TODO: test attachments with different content-type, encoding, and filename
         response = self.submit_form(
             data={
                 "from": "john@doe.com",
@@ -143,23 +158,34 @@ class TestMailStore(unittest.TestCase):
                 ],
                 "subject": "test subject",
                 "block_id": "form-id",
+                "attachments": {
+                    "file_1": {
+                        "field_id": "file",
+                        "encoding": "base64",
+                        "data": base64.b64encode(file_str),
+                        "filename": "file.pdf",
+                        "content-type": "application/pdf",
+                    },
+                },
             },
         )
-        transaction.commit()
-
         self.assertEqual(response.status_code, 200)
+
         response = self.export_data()
         data = response.json()
         self.assertEqual(len(data["items"]), 1)
         self.assertEqual(
             sorted(data["items"][0].keys()),
-            ["__expired", "block_id", "date", "id", "message", "name"],
+            ["__expired", "block_id", "date", "file", "id", "message", "name"],
         )
         self.assertEqual(
             data["items"][0]["message"],
-            {"label": "Message", "value": "just want to say hi"},
+            {"label": "Message", "value": "just want to say hi", "field_type": "text"},
         )
-        self.assertEqual(data["items"][0]["name"], {"label": "Name", "value": "John"})
+        self.assertEqual(
+            data["items"][0]["name"],
+            {"label": "Name", "value": "John", "field_type": "text"},
+        )
         response = self.submit_form(
             data={
                 "from": "sally@doe.com",
@@ -182,13 +208,35 @@ class TestMailStore(unittest.TestCase):
         )
         self.assertEqual(
             sorted(data["items"][1].keys()),
-            ["__expired", "block_id", "date", "id", "message", "name"],
+            ["__expired", "block_id", "date", "file", "id", "message", "name"],
         )
-        sorted_data = sorted(data["items"], key=lambda x: x["name"]["value"])
-        self.assertEqual(sorted_data[0]["name"]["value"], "John")
-        self.assertEqual(sorted_data[0]["message"]["value"], "just want to say hi")
-        self.assertEqual(sorted_data[1]["name"]["value"], "Sally")
-        self.assertEqual(sorted_data[1]["message"]["value"], "bye")
+        # sorted_data = sorted(data["items"], key=lambda x: x["name"]["value"])
+        self.assertEqual(data["items"][0]["name"]["value"], "Sally")
+        self.assertEqual(data["items"][0]["message"]["value"], "bye")
+
+        self.assertEqual(data["items"][1]["name"]["value"], "John")
+        self.assertEqual(data["items"][1]["message"]["value"], "just want to say hi")
+
+        self.assertEqual(
+            data["items"][1]["file"]["value"]["contentType"], "application/pdf"
+        )
+        self.assertEqual(data["items"][1]["file"]["value"]["filename"], "file.pdf")
+        self.assertEqual(data["items"][1]["file"]["value"]["size"], 74429)
+        self.assertRegex(
+            data["items"][1]["file"]["value"]["url"],
+            r"/example-context/saved_data/@@download/[0-9]+/file/file.pdf$",
+        )
+
+        # get file
+        self.browser.addHeader(
+            "Authorization",
+            f"Basic {SITE_OWNER_NAME}:{SITE_OWNER_PASSWORD}",
+        )
+        self.browser.open(
+            data["items"][1]["file"]["value"]["url"],
+        )
+        self.assertEqual(self.browser._response.status, "200 OK")
+        self.assertEqual(self.browser.contents, file_str)
 
         # clear data
         response = self.clear_data()